Typechecking of SIMD vector field accessors
This produces the 'correct' typecheck result for SIMD vector field
accesses, but does not implement them yet. To make this work we had to
turn the n-expression of the simd!(…) syntax extension into a literal,
unfortuantely, but this shouldn't be much of an issue, in retrospect,
you probably don't want to allow that anyhow.