Adds support for llvm! syntax extension
This is the most stupid syntax extension around, it allows you to break
absolutely everything. Bad Apple!! style everything.

Don't think that Rust will come and save you from the eternal darkness
that may or may not be brought upon you by using this syntax extension.
It adds _RAW_ LLVM IR to your program, and this means you will have to
check the code extra carefully.

This is an example, which reads the LLVM cyclecounter.

    llvm!(extern {
      ir "declare i64 @llvm.readcyclecounter()";

      fn readcyclecounter() -> i64 {
        "%result = call i64 @llvm.readcyclecounter()
         ret i64 %result"
      }
    })